requirements.txt
Flask
openmeteo-requests
opencage
requests-cache
retry-requests
numpy
pandas
python-dotenv